Stop section assembly from throwing away the form clustering found
Phase 2 of the pipeline audit (PIPELINE-AUDIT.md, new — it tracks all four
phases and their evidence). The two high-severity findings here are the same
shape as Phase 1's downbeat bug: an earlier stage gets the right answer and a
later stage discards it. Both were reproduced before being fixed.

  a short intro destroyed the song's form

Reproduced on the most common shape in the repertoire — 2-bar intro + verse ×4.
Clustering got it exactly right: group B, four occurrences, at bars 2/6/10/14.
Then `_absorb_runts` merged the runt intro forwards, handed the merged section
the *intro's* group, and flattened `repeats: 4`. The song came out as ONE
eighteen-bar section labelled with the intro's letter — so `compile` attached
the intro's strum pattern (two bars of evidence, in practice the fallback) to
sixteen bars of verse, the verse group's pooled pattern was computed and
referenced by nothing, and the rail showed one undifferentiated blob.

Absorption now always keeps the *host's* group, and never expands a host with
`repeats > 1`. Where the floor would cost a collapsed repeat, the runt stays a
short section of its own instead. That relaxes §15's ~4-bar floor, deliberately:
the floor is a preference, `repeats` is an encoding both the container and the
rail read, and lint imposes no minimum length. A standalone two-bar intro is a
far smaller lie than an eighteen-bar section claiming to be one piece of music.

  the last section of most songs left its own repeat group

`_chunk` merged a short tail into the last block *before* clustering. That block
then had the wrong length, and `block_similarity` scores unequal lengths 0 — so
verse ×4 + a 2-bar tag gave group A occurrences [0, 4, 8] and made the final
verse a group of its own: no consensus correction, onsets never pooled into the
verse's pattern, drawn on the rail as different music. Since songs rarely end on
an exact multiple of their own period, this was most songs. §15's
tail-joins-previous rule is right at section level and wrong before clustering,
so it moved: the tail is its own block, and the floor is applied afterwards.

  the twelve-bar blues was not in the list of periods

`CANDIDATE_UNITS` had no 12, so core repertoire was chopped at period 4 and its
I/IV/V phrases scattered across groups that are not sections — three choruses
came out `A B B A B B A B B`. Added 12, and 6 for six-bar phrases and the 12/8
blues.

  and two smaller ones

The labelling vocabulary was unfinished: `preChorus` and `bridge` are in the
container's SECTION_KINDS and were never emitted, and `harmony.is_dominant_of`
was written for the pre-chorus half-cadence cue, called by nothing, with a
docstring pointing at a `form.label` that does not exist. Now a unique mid-song
group is a `bridge`, and a repeated non-chorus group ending on V immediately
before the chorus is a `preChorus` — gated on the song having two or more
repeated non-chorus groups, because a lone verse ending on V to lead into the
chorus is the most ordinary thing in this repertoire and would otherwise be
relabelled everywhere. `solo` stays unreachable and now says why: it is a
question about timbre.

`EnergyCurve.hop_ms` was rounded to a whole millisecond (46.44 → 46), which is a
~1% *rate* error, not a rounding error. By the fourth minute the window read for
a bar sat two seconds — most of a bar — from the bar it claimed to measure,
degrading the one comparison the probe exists for. It is a float now.

  measured

`--theory`, 15 tracks, against HEAD~1:

  truth  MEAN   0.963 -> 0.963   the no-op on perfect input still holds exactly
  btc    MEAN   0.825 -> 0.825   19 bars rewritten, was 16

  here-comes-the-sun  0.788 -> 0.791    sects 14 -> 15
  ob-la-di            0.845 -> 0.847    sects 16 -> 18
  twist-and-shout     0.896 -> 0.902    sects  3 ->  5
  let-it-be           0.807 -> 0.792    sects  4 ->  7

Delivered-neutral on accuracy, and consensus-off is identical everywhere, which
is the expected result since none of this touches chord recognition. What moved
is the structure: every affected track gained sections, because intros and tags
are no longer swallowed, and the vote is now taken over groups that really are
the same music. Three tracks up, Let It Be down 0.015. Reported rather than
buried — as with §20, the coherence is what was bought, except that here the
coherence *is* the deliverable, since the rail and the per-group strum pattern
are what the player reads.

One change was tried and reverted, recorded because it looks obviously right:
scoring the phase search on full-length blocks only. A runt matches nothing, so
excluding it seems like removing noise — but it lets the search shift the grid
until the song's one misheard bar lands in the tail runt and then score a
perfect 1.0 on what is left. test_consensus.py caught it. The runt's 0 is real
information about the offset.

Tests 423 -> 432. The old intro test asserted only on *groups*, which is why it
stayed green through the whole of B1; there is now one asserting sections.

+++ app/analysis/types.py
@@ -101,7 +101,13 @@ class EnergyCurve:
     one.
     """
 
-    hop_ms: int
+    # **Float, and it has to be.** A hop is 1024 samples at 22.05 kHz, i.e.
+    # 46.44 ms, and rounding that to 46 is not a rounding error — it is a 0.95%
+    # *rate* error that accumulates. Three minutes in, the window read for a bar
+    # sat two seconds — most of a bar — away from the bar it claimed to measure,
+    # so the verse/chorus comparison this curve exists to make was quietly being
+    # taken over the wrong music.
+    hop_ms: float
     values: list[float]
 
     def mean_between(self, start_ms: float, end_ms: float) -> float:
